Move workspace test
