+++ tests/workspace_test/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "clippy_workspace_tests"
+name = "workspace_test"
 version = "0.1.0"
 edition = "2018"
