+++ tests/workspace_test/path_dep/src/lib.rs
